Debugging post processing issue
Random flicker occurs now that post processing system has been implemented, seems to be down to certain swapchain images being presented (image 2?) rather than post processing effect itself
TODO:
- Create some logic for forcing engine fps instead of running as fast as possible(?)
- Look into RenderDoc C++ API(?)

+++ Cassidy/Utils/GlobalTimer.cpp
@@ -5,7 +5,7 @@
 void GlobalTimer::updateGlobalTimerImpl()
 {
   uint64_t currentTimeMS = SDL_GetTicks64();
-  const double currentTimeSecs = currentTimeMS * 0.001;
+  const double currentTimeSecs = static_cast<double>(currentTimeMS) * 0.001;
 
   m_deltaTimeSecs = static_cast<float>(currentTimeSecs - m_engineTimeSecs);
   m_engineTimeSecs = currentTimeSecs;
